+++ MANIFEST.in
@@ -1,4 +1,5 @@
 include dash_bio/bundle.js
 include dash_bio/metadata.json
 include dash_bio/package.json
+include dash_bio/genomeviewer-style.css
 include README.md
