+++ src/lib/index.js
@@ -7,7 +7,7 @@ import Ideogram from './components/Ideogram.react';
 import Circos from './components/Circos.react';
 import AlignmentChart from './components/AlignmentChart.react';
 import AlignmentViewer from './components/AlignmentViewer.react';
-import GenomeViewer from './components/GenomeViewer';
+import GenomeViewer from './components/GenomeViewer.react';
 
 export {
     Speck,
